certificate-sniffer/requirements.txt
requests==2.25.1
cryptography==3.4.7
validators==0.18.2
pyOpenSSL==20.0.1
python_dateutil==2.8.1
